Split Expr type file
