+++ src/Language/Mimsa/Typechecker/DataTypes.hs
@@ -8,8 +8,8 @@ where
 
 import Data.Map (Map)
 import qualified Data.Map as M
-import Language.Mimsa.Types.AST
 import Language.Mimsa.Types.Construct
+import Language.Mimsa.Types.DataType
 import Language.Mimsa.Types.Environment
 import Language.Mimsa.Types.MonoType
