+++ src/Language/Mimsa/Types.hs
@@ -3,7 +3,8 @@ module Language.Mimsa.Types
     module Language.Mimsa.Types.TypeConstructor,
     module Language.Mimsa.Types.ExprHash,
     module Language.Mimsa.Types.Usage,
-    module Language.Mimsa.Types.AST,
+    module Language.Mimsa.Types.Expr,
+    module Language.Mimsa.Types.Literal,
     module Language.Mimsa.Types.Store,
     module Language.Mimsa.Types.TypeError,
     module Language.Mimsa.Types.MonoType,
@@ -26,18 +27,24 @@ module Language.Mimsa.Types
     module Language.Mimsa.Types.StoreExpression,
     module Language.Mimsa.Types.Project,
     module Language.Mimsa.Types.TypeName,
+    module Language.Mimsa.Types.FuncName,
+    module Language.Mimsa.Types.StringType,
+    module Language.Mimsa.Types.DataType,
   )
 where
 
-import Language.Mimsa.Types.AST
 import Language.Mimsa.Types.Bindings
 import Language.Mimsa.Types.Construct
+import Language.Mimsa.Types.DataType
 import Language.Mimsa.Types.Environment
 import Language.Mimsa.Types.Error
+import Language.Mimsa.Types.Expr
 import Language.Mimsa.Types.ExprHash
 import Language.Mimsa.Types.ForeignFunc
+import Language.Mimsa.Types.FuncName
 import Language.Mimsa.Types.InterpreterError
 import Language.Mimsa.Types.Library
+import Language.Mimsa.Types.Literal
 import Language.Mimsa.Types.MonoType
 import Language.Mimsa.Types.Name
 import Language.Mimsa.Types.Project
@@ -47,6 +54,7 @@ import Language.Mimsa.Types.Scheme
 import Language.Mimsa.Types.Scope
 import Language.Mimsa.Types.Store
 import Language.Mimsa.Types.StoreExpression
+import Language.Mimsa.Types.StringType
 import Language.Mimsa.Types.Substitutions
 import Language.Mimsa.Types.Swaps
 import Language.Mimsa.Types.Tui
