+++ src/Language/Mimsa/Types/Environment.hs
@@ -6,8 +6,8 @@ import Data.Map (Map)
 import qualified Data.Map as M
 import qualified Data.Text as T
 import Language.Mimsa.Printer
-import Language.Mimsa.Types.AST
 import Language.Mimsa.Types.Construct
+import Language.Mimsa.Types.DataType
 import Language.Mimsa.Types.Scheme
 import Language.Mimsa.Types.Variable
