+++ src/Language/Mimsa/Types/InterpreterError.hs
@@ -6,7 +6,7 @@ import Data.Map (Map)
 import qualified Data.Map as M
 import qualified Data.Text as T
 import Language.Mimsa.Printer
-import Language.Mimsa.Types.AST
+import Language.Mimsa.Types.Expr
 import Language.Mimsa.Types.Name
 import Language.Mimsa.Types.Scope
 import Language.Mimsa.Types.Variable
