+++ src/Language/Mimsa/Types/Name.hs
@@ -10,7 +10,6 @@ import qualified Data.Char as Ch
 import Data.Text (Text)
 import qualified Data.Text as T
 import GHC.Generics
-import Language.Mimsa.Printer
 
 newtype Name = Name {getName' :: Text}
   deriving stock (Eq, Ord, Generic)
@@ -21,9 +20,6 @@ newtype Name = Name {getName' :: Text}
       JSON.ToJSON,
       JSON.ToJSONKey
     )
-
-instance Printer Name where
-  prettyPrint = getName
 
 getName :: Name -> Text
 getName (Name t) = t
