+++ src/Language/Mimsa/Types/StringType.hs
@@ -1,0 +1,13 @@
+{-# LANGUAGE DerivingStrategies #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
+
+module Language.Mimsa.Types.StringType
+  ( StringType (..),
+  )
+where
+
+import qualified Data.Aeson as JSON
+import Data.Text (Text)
+
+newtype StringType = StringType Text
+  deriving newtype (Eq, Ord, Show, JSON.FromJSON, JSON.ToJSON)
