+++ src/Language/Mimsa/Types/TypeError.hs
@@ -13,9 +13,10 @@ import qualified Data.Set as S
 import Data.Text (Text)
 import qualified Data.Text as T
 import Language.Mimsa.Printer
-import Language.Mimsa.Types.AST
 import Language.Mimsa.Types.Construct
+import Language.Mimsa.Types.DataType
 import Language.Mimsa.Types.Environment
+import Language.Mimsa.Types.Expr
 import Language.Mimsa.Types.MonoType
 import Language.Mimsa.Types.Name
 import Language.Mimsa.Types.Swaps
